Make definition of vanessa_socket_pipe_read_write_func match its declaration
Resolves the following build problem on Debian Sid kfreebsd-amd64

vanessa_socket_pipe.c:245: error: conflicting types for 'vanessa_socket_pipe_read_write_func'
vanessa_socket.h:423: error: previous declaration of 'vanessa_socket_pipe_read_write_func' was here

+++ libvanessa_socket/vanessa_socket.h
@@ -420,7 +420,7 @@ int vanessa_socket_pipe_func(int rfd_a,
  *         -1 on error
  **********************************************************************/
 
-int vanessa_socket_pipe_read_write_func(int rfd,
+ssize_t vanessa_socket_pipe_read_write_func(int rfd,
 					int wfd,
 					unsigned char *buffer,
 					int buffer_length,
